refactor: move the registration scanner into djls-project
