+++ crates/djls-semantic/src/python/analysis/exceptions.rs
@@ -8,7 +8,7 @@ use ruff_python_ast::StmtRaise;
 use crate::python::analysis::expressions::eval_expr;
 use crate::python::analysis::state::AbstractValue;
 use crate::python::analysis::state::Env;
-use crate::python::ext::ExprExt;
+use djls_project::extraction::ExprExt;
 use crate::python::types::ExtractedMessageArg;
 use crate::python::types::ExtractedMessageTemplate;
