+++ crates/djls-semantic/src/python/blocks/opaque.rs
@@ -7,7 +7,7 @@ use ruff_python_ast::statement_visitor::StatementVisitor;
 use ruff_python_ast::statement_visitor::walk_stmt;
 
 use crate::python::blocks::is_parser_receiver;
-use crate::python::ext::ExprExt;
+use djls_project::extraction::ExprExt;
 use crate::python::types::BlockSpec;
 
 /// Detect opaque block patterns: `parser.skip_past("endtag")`.
